switch to common setup approach
