+++ packages/python/test/integration-1.test.ts
@@ -1,15 +1,14 @@
-const fs = require('fs');
-const path = require('path');
+require('./integration-setup')(1);
+
+// NOTE: The fixture `00-requrest-path` has special case handling, which is why
+// it's broken out below. This is also why this fixture is filtered out of the
+// chunking in "integration-setup".
+
 const assert = require('assert');
 const fetch = require('node-fetch');
 const execa = require('execa');
 const { spawn } = require('child_process');
-
-const {
-  testDeployment,
-} = require('../../../test/lib/deployment/test-deployment.js');
-
-jest.setTimeout(4 * 60 * 1000);
+const path = require('path');
 
 const fixturesPath = path.resolve(__dirname, 'fixtures');
 
@@ -70,16 +69,3 @@ it('should match the probes against Python dev servers', async () => {
     process.kill(asgiServer.pid);
   }
 });
-
-const allFixtures = fs.readdirSync(fixturesPath);
-const chunkFixtures = allFixtures.slice(0, 10);
-
-// eslint-disable-next-line no-restricted-syntax
-for (const fixture of chunkFixtures) {
-  // eslint-disable-next-line no-loop-func
-  it(`should build ${fixture}`, async () => {
-    await expect(
-      testDeployment(path.join(fixturesPath, fixture))
-    ).resolves.toBeDefined();
-  });
-}
